LAI and needles length made automatic.
Xianglin section updated with an autotomatic max needle parameter based on the culmative sum of the needle growth from last year and the current needle growth.

The outputs of each function are also not pushed to a vector after their respective functions.

Preles second year indexes corrected for GPP etc.