A new strgae function added based on the biomass of each organ. This is based on the mass of each organ so this is hard coded. The relative size of this mass is then used to split the xylem and phloem processes more fairly.
As the stroage components are a nice way of representing the "need" of the organ I baed the movement of the sugar on this and updated the equations accordingly with also an upper limit in case the tree runs out of stroage.

A lower limit has been added to all of the transfers indervidually and then the sugar as a whole to represente that fact that the enzyme equation has to be negative, but sugar can't be so it is first applied and then sugar is checked for negatvity. The same thing for starch.

Then some capitals were changed